Update error summary to be able to handle nested object errors
At the moment, when nested objects have errors, the attribute
of the error is constructed via chaining the associations together.
this.

A consultation_response_form_data error for title would return this
as the attr:

"edition.consultation_response_form.consultation_response_form_data_title"

As there is an expectation that the the id is constructed with we need
to gsbug the "."'s to "_"'s'