Add migration to split WorldLocation into two parts
This splits the existing `WorldLocation` records into two parts:
`WorldLocation` (which will store details of the location itself) and
`WorldLocationNews` (which will store the details needed to render the
news page).

We are using raw SQL in the migrations as these tables are translatable,
meaning Rails will return only some translations and not others if we
refer to the model by it's name. Using a `SELECT` statement will ensure
we get all translations.

The migration is irreversible as once the migration has been run, it
will be almost impossible to bring the translations back into sync with
each other.