Republish Topical Event when features are reordered
When a topical event's features are reordered, we need the topical event
to be republished so the content item includes the updated order of
features.

Whilst there is a callback on the `TopicalEventFeaturing` model to
republish the associated topical event, this was not being called as the
`order` method uses `update_column` which skips callbacks.

Changing this to `update` would not be desirable, as the topical event
will get republished once for each featured item (which could be
multiple items).  Therefore adding a single republish request so we get
the desired behaviour of the topical event only being republished once.