Add Duplicate Edition Controller Test Helpers
These are just duplicates of the legacy files with the "Preview design
system" permission added to the current user.

As the edit view is amended in the GOV.UK Design System these tests
will begin to fail as the UI changes.

This should allow us to keep full test coverage and simply delete the
legacy files once we've shipped the new/edit pages.