Add enum for world_location_type
We are currently using the `world_location_type_id`  column to act as a
key for the entries in the WorldLocationType model, which only contains
2 entries and behaves exactly like an enum.

Therefore we can actually use an enum here, which should allow us to
reduce a lot of code complexity.

The key numbering is maintained in line with the existing
WorldLocationType entries.

This also automatically gives us new helper methods on the WorldLocation
model: `world_location_types`, `world_location?` and
`international_delegation?`.