Update usage of all_by_type method for world locations
This method was previously using a sort order that was defined in the
`WorldLocationType` model and outputting a tuple.

Now that model has been removed, we need to define the sort order
differently.  In this case, we are now using the `in_order_of` method
that was introduced in Rails 7.