Normalize locale files
This uses the `translate:normalize` rake task to normalize the locale
files in accordance with our predefined rules.

As this hasn't been done when the locale files were recently changed, we
ended up with unrelated changes when importing new translations in
another PR.

Therefore normalizing separately so we don't get these unrelated
changes.