Republish WorldLocationNews when feature changes
When a featured document changes, we need to republish the
WorldLocationNews content to make this change live on the website.

As these pages were previously rendered by Whitehall, it did not matter
that this didn't happen.  But now we render these pages in Collections
using the content item, this update is essential.