Add similar slug warning to error summary component
If an edition has the same title as another document then we
show the user a warning which lets them know that they need to
update the title.

After some discussion with the design team, we've decided that
presenting this as an error in the error summary is preferable to
having multiple banners.