Update information about frontend mode in README
The README previously suggested the content rendered by Whitehall was
legacy content, implying the content is no longer current. That is not
true as these content types are regularly updated.

Chaning the text so it explains what is meant by the term "legacy", i.e.
not yet migrated to Publishing API.