Remove logic to allow any offsite link on HM Queen Elizabeth II's page
This reverts commit 39d4fb8a2fafbb72ccf3c8077c70d34c9bdbaf70.

It was a temporary special case during the time leading to HM Queen Elizabeth
II's funeral.  It adds royal.uk to the allowlist as it's still featured and it
was previously approved by the Product Owner. We are not expecting other
offsite links to be added to this topical event page.