Fix display_type for languages which don't have singular nouns
We were previously losing translations and defaulting to English for the edition display type for languages which don't tend to have singular nouns, e.g. Mandarin, since we were specifically looking for the translation for 'one', whereas for some languages only a translation for 'other' exists. This change uses I18n's internal logic for deciding which translation (/) to look for, depending on the locale, so we shouldn't be losing translations for display types now.