Ensure Javascript does not clear person_override field
At the moment, the EditionPublishing JS module looks for the
person_override_active checkbox. If it finds it and it's checked it
toggles a text area, if not then it clears the value in the text area
and hides it.

We don't want this behaviour to occur in the new implementation.
I've update the module to check for the presence of the checkbox before
clearing the text input.