Fix bug in error summary which causes `Error#full_message` to blow up
At the moment, we're passing in an object as it's parent object
to allow us to link errors to the input which generally uses
the parent class as it uses shared form fields across all edition and
attachment types.

However, this blows up when `Error#full_message` is called on an
error in which a module is included on a child class but not
in the parent class.

This occurred when `.role_appointment` was called on the object passed
in which was an Edition, rather than a Speech.

By always iterating through the actual objects errors, and not the
object masquerading as it's parent class this won't occur.

However, we still need a way to link errors to the input, so i've
added `parent_class` as an optional param.