Ensure external url JS doesn't action when page is using GOV.UK Design System
In the Bootstrap implementation, JS looks for a checkbox and clears the
value of the external url field if the checkbox is not checked.

The ID of the checkbox is slightly different when you use the publishing
component so it's not seeing that it's checked.

We no longer need this functionality so i'm only running the function
when it can find the Bootstrap checkbox.