Clean external url params
When the `external` checkbox is unchecked. The external_url field
needs to be set to nil before the edition is persisted.

We also need to pass a value through as a hidden field to ensure
`external` is set to false. Currently, the publishing component checkbox
doesn't handle this and pass a value when unchecked.