Remove duplicate error messages from the Edit Edition page
At the moment, when an object is passed to the draft updater to
see if it can be passed downstream, it checks to see if the edition
is in a draft state and if so, whether the object is valid.

```
def failure_reason
  if !edition.pre_publication?
    "A #{edition.state} edition may not be updated."
  elsif !edition.valid?
    "This edition is invalid: #{edition.errors.full_messages.to_sentence}"
  end
end
```

This is then set as a variable in the controller

```
@information = updater.failure_reason
```

and then passed into the `standard_edition_form` method and concatonated
into the template

```
concat edition_information(information) if information
```

As we now handle errors in via the summary component, we no longer
need this functionality.

Originally i thought that there might be an issue with the
"A #{edition.state} edition may not be updated." not showing, but
this before_action handles it anyway and redirects to the summary
page.

```
before_action :prevent_modification_of_unmodifiable_edition, only: %i[edit update]
```

```
def prevent_modification_of_unmodifiable_edition
  if @edition.unmodifiable?
    notice = "You cannot modify a #{@edition.state} #{@edition.type.titleize}"
    redirect_to admin_edition_path(@edition), notice:
  end
end
```

This means we can just remove the @information functionality all
together in the GOV.UK Design System implementation