Add design system edit view under a feature flag
This commit renames the edit view to edit_legacy and creates a new edit view using design system modules.
It also updates the NeedsController to allow both of these views to co-exist whilst putting the design
system view under a feature flag.

Finally, it adds a feature test for the new design system edit view. One of the steps was renamed to account
for the new reworded button name.