Add nation_fields partial in the GOV.UK Design System
This adds in a checkbox field with conditionally revealed external_url
inputs which are rendered for Publications, Consultations and Detailed
Guides.

Unfortunately it's required a pretty substantial amount on munging to
wire it into the controller.

There's also an issue in that the controller munging needs to be done
prior to the edition being built in the `build_edition` before action
so the correct attrs are assigned.

However, as the edition hasn't been built you're unable to see what
the class of the object is.

I've gotten round this by checking which controller i'm in but it seems
a little hacky. There may well be a better way.