Add support for upcoming changes section
We need to add an 'Upcoming changes' section to the 'What's new' page in Whitehall Publisher.

This section would be between the introduction section and the 'recent changes' section.

The section and entries follow the same format as the 'recent changes' section.

The section shouldn't show in the contents list or on the page if there's no entries/content in the YAML file.

We're using the page to communicate with publishers upcoming changes to Whitehall Publisher.

https://github.com/alphagov/whitehall/blob/main/config/locales/en/admin/whats_new.yml

https://github.com/alphagov/whitehall/tree/main/app/views/admin/whats_new