Move historic appointments routes into a scope
This simplifies the routes for historic appointments by removing a lot
of the unusual constraints and replacing them with hardcoded routes.

It appears there was part completed work done to make the Historic
Appointments controller more generic, but was never done. Unrolling that
work here to make the routes easier to understand.

Hardcoding a value in the controller to avoid needing the constraints,
as the `index` method is solely used by `past-prime-ministers`.

Also replacing `person_id` with `id` to make the values consistent
between the past foreign secretaries and past prime minister pages.