Update UrlMaker tests for non-localised atom feed
These were using a content type that is no longer rendered by Whitehall.
This meant we needed to maintain an entry in the routes file for this
content type, despite the route no longer being needed.

Removing the test as Whitehall no longer renders any non-localised
content as atom feeds.