Explicitly form atom feed URL for Topical Events
We need to explicitly use the correct helper method (and pass in the
correct parameters) for the Topical Event's atom feed, as we are no
longer using the helpers defined in the routes file.

This change can be removed when the email signup code
(https://github.com/alphagov/whitehall/pull/6900) has been deleted as
that was the only code requiring this method.