Add test for Get Involved admin page
There is currently no test that loads the admin page for Get Involved.

This meant that an error in loading the page (e.g. by breaking a route
helper) would not be noticed by running the test suite.

I'm not exactly sure what this page is meant to do, other than provide
some links to other places, so just adding a basic test that the page
loads without error (which would be enough to identify a broken route
helper, for example).