Implement previewable text area for GovSpeak
We need a component that can be used for Govspeak textarea input fields.

It must include an 'inline preview' feature so that users can preview their Govspeak/Markdown rendered as HTML.

So publishers can preview Govspeak content before saving/publishing it. We know that this is an important feature for publishers.

https://trello.com/c/FAig6Fn0/792-implement-previewable-text-area-for-govspeak