Fix flakey scheduled publishing test
When running tests in parallell the scheduled publishing worker test has
a tendency to randomly fail. This seems to be because old editions are
not properly discarded or are running at the same time which causes the
test to fail.

To fix this we have have added a method to pass in a different worker
name to the test sidekiq, this means that the editions should no longer
overlap since they are in different queues.

Trello card: https://trello.com/c/DhRIp7Nu/830-flaky-scheduled-publishing-worker-tests-in-whitehall

Co-authored-by: BeckaL <becka.lelew@digital.cabinet-office.gov.uk >