Update links to publication finder
The 'How Government Works' page uses route helpers for the publications
finder, which is no longer rendered by Whitehall and currently redirects
to an alternate location rendered by Finder Frontend.

This replaces the URL with a link to the final destination, to avoid the
redirect and avoid the use of a route helper.

Note: the diff on the new lines is because there were previously blocks
of spaces that have been automatically removed by my IDE.