Update route helpers for feeds
These feeds are no longer rendered by Whitehall, however we currently
use a Whitehall route helper to generate the URL, which then redirects
to a Finder Frontend rendered path.

Updating these helpers allows us to link directly to the finder without
needing the redirect or route helper.

Note the query parameters are different between the Whitehall and Finder
Frontend finders, so this has been updated here (it was previously done
by the redirect).