Update consultations redirect
This redirect currently uses the publications route helper, which has
now been removed.

The previously returned route also issued a redirect to another path, so
updating the first redirect to now point to the final destination of the
original redirect.

This allows us to remove usage of the route helper method.