Move logic from image_template in a View Component
This pulls out the iterative logic from the image_template into a
View Component and unit tests it.

There's quite a difference in logic based on various states, such as
the image being persisted or new, and whether a file cache is present.

Unit testing all these permutations seems prudent.