Remove redundant step from Fatality Notice cucumber tests
This test contained a step `I can navigate to the list of announcements`
that appears to have been leftover from the times when Whitehall
rendered organisation pages.

All it does it stubs the organisation page, but that page never gets
loaded in these tests, so running this code is redundant.

The step was previously used in the announcement finder tests, but they
have now been removed, so the step is safe to delete.