Standardise feature tests for Design System pages
There are various step definitions that need to perform different actions depending on whether they're on a Design System page or a Bootstrap (legacy) page in the Whitehall admin. For example, the steps involved in publishing a new document differ slightly between the two design systems.

A couple of different approaches were used to detect which type of page we're on. These included:

- `@user.can_preview_design_system?`
- `@user.has_permission? "Preview design system"`

However, these depend on the `@user` object (i.e. the currently logged in user) and make the assumption that users with the "Preview design system" permission will be shown a Design System page, and those without will be shown the legacy Bootstrap interface. But ideally feature tests would use the application as if they were a real-life human user, without being able to 'see behind the scenes' and inspect objects directly.

To rationalise these approaches and remove their coupling to the `@user` object, I've introduced a new helper method called `using_design_system?`. This helper inspects the web page itself to work out if we're on a Design System or Bootstrap page.

It'll return `true` if the HTML page contains:
```
<html class="govuk-template">
```

This is the class name required to add Design System styling to the page.

Otherwise it'll return `false`, meaning the page isn't using the Design System layout.