Remove `publications_filter_path` route helper
This was only being used in one view (which will soon be migrated out of
Whitehall anyway), so it seems reasonable to remove the helper.

The link for 'Corporate Reports' was previously broken and linked to a
finder containing all publications for the organisation. This has been
somewhat improved here by linking to a subset of publications.