Normalize locale files
Ongoing work to remove finders involve updating locale files in bulk,
which requires us to run `rake translations:normalize`.

As the locale files are not currently normalized, this will result in
this diff being added each time we run this.

Updating them now before we start manipulating the files in bulk.