Add ability to show or hide banner from yaml file
We've discovered a need to toggle showing or hiding the whats new
banner between releases.

The simplest implementation of this is to just add a value to the
YAML file that a Content Designer can set to true or false depending
on whether they want it to render or not.

Testing this is a bit strange. After some discussion it feels like the
simplest thing to do is to conditionally test whether the banner is
rendered or not based on the current state of the whats_new.yml.

If show_banner is set to true it checks for the presence of the banner,
if false it checks for the abscence.