Add margin to navbar for admin template when Whats new banner is absent
The design system and admin layout use a shared header. When it was
merged into a shared file and the 'Whats new' banner was added,
the top-margin was removed as the banner provides margin.

Now when the banner is removed between releases, we no longer have that
margin.

This adds margin to admin views, when the banner is not present.