Add temporary rake task to unpublish non-English finders
This rake task will need to be run in production (then deleted) and
serves the purpose of marking the non-English finders as 'gone' (410
response code).

Currently the locale versions of the finders do not exist and are
redirected at the router level to `/government`, which is then handled
by Whitehall.

Therefore we need to first publish the content item at the path for the
translation before we can mark it as gone.