Revert design of command paper and parliamentary_session to use checkboxes
We're reverting this to use checkboxes and rather than radios so it's
closer to the original implementation.

As the command_paper_number and house_of_common number fields sit under
a h2, i've updated the labels of the fields to be one size smaller
then the heading_size of the rest of the fields.