Add field container helper
At the moment, errors in the error summary component link directly
to the inout. This means the label and hint text are not
visible on the screen and a user loses context.

Instead of linking directly to the input, we're going to wrap the input
in a div with the current input id. We will then append the id of the
input with '-field'.

This helper takes an id and a block. It creates a div with the id passed in
and renders the block within it.