Switch feature tests to 'Preview design system'
Previously the feature tests would apply the 'Preview next release'
flag, introduced in PR #7001.

But as outlined in that PR description, that was only a first step.
Ideally we should be testing against the 'Preview design system' flag
so we cover all Design System changes, not only those flagged for the
next release.

I've also changed the `using_design_system?` method to detect whether
we're running in Design System mode by reading from the environment
variable `CUCUMBER_PREVIEW_DESIGN_SYSTEM`. This makes it more resilient
to regressions where a Bootstrap layout is incorrectly rendered when
should have rendered the Design System, because it'll no longer sniff
the page itself for presence of Design System page elements.