Introduce Design System tags for feature tests
This commit adds support for three tags on Cucumber feature tests:

1. `@design-system-wip`
=======================

Denotes a feature which is still a work-in-progress in the Design System
preview. Tests for this feature are known to be failing because it's
still being worked on.

Features with this tag will be run against Bootstrap (the 'default'
profile) but not against the Design System (the 'preview_design_system'
profile).

2. `@bootstrap-only`
====================

Denotes a feature which is specific to the legacy Bootstrap admin
layout.

Use this when the Design System version of a feature has changed
considerably, such that it now warrants a separate feature definition.

Similar to `@design-system-wip`, features with the `@bootstrap-only` tag
will be run against Bootstrap but not the Design System.

3. `@design-system-only`
========================

The opposite of `@bootstrap-only`. This denotes a feature which is
specific to the new Design System admin layout.

Features with this tag will only run against the Design System preview.

In most cases, `@design-system-only` features should be accompanied by a
`@bootstrap-only` feature to cover the equivalent feature. This allows
the example steps (i.e. the user journey) to differ between Bootstrap
and Design System versions of a feature.