Resend child and parent organisations of a reslugged org to search
Since the search data for an organisation contains slugs for parent and child organisations, when an organisation is reslugged, its child and parent organisation entries in search can be out of sync, as they will contain the old slug for the reslugged organisation. This has caused issues for us since the organisations api uses search data, and so when this is out of sync it can cause problems in other parts of our stack.