Remove LockedDocumentConcern
This concern was added at a time where the was an expectation that
content would be migrated from Whitehall to Content Publisher in the
near future [1].

We're now over 3 years later and it looks very unlikely that this
migration will be prioritised again. We can therefore remove this code
as it is redundant.

[1]: https://github.com/alphagov/whitehall/pull/4930