Remove locked field from Document
Following from the previous commits that removed the usage of a `locked`
field this ignores the field on the ActiveRecord model. This is a
precursor to the field being dropped from the database.

I've chosen to do the migration that removes the field in a separate PR
so that this can be rolled back if there are any problems.