Present worldwide organisation links for International Delegations
At the bottom of International Delegation pages is a section that
includes the summary page of the associated Worldwide Organisation.

This text is an editionable piece of content and is stored as a
Corporate Information Page associated with the organisation.

We therefore need to include it as a link in order to render the summary
content on the international delegation page.