Fix failing Design System feature specs
This commit fixes feature specs that were failing when the 'Preview
design system' flag was enabled.

This included adapting to changes in form label text between Boostrap
and the Design System layouts.

I've also added a new spec helper method called
`within_conditional_reveal`. It behaves similarly to the `within` helper
built in to Capybara, which accepts a block and will scope all `find`
calls to within the specified parent container.

In this case, it can be used to scope `find` calls to the contexts of a
conditional reveal, identified by its label text.

For example, it could be used like this:

```ruby
choose "Email"
within_conditional_reveal "Email" do
  fill_in "Email address", with: "me@example.com"
end
```

In this example, the `fill_in` command will populate the email address
field within the conditional reveal.