Fix formatting inconsistencies in feature specs
The files `features/*.feature` contained various formatting and
indentation inconsistencies.

This is starting to cause me problems because my code editor (Visual
Studio Code) insists on reformatting these files as I work on them.

The formatting that it applies seems to be consistent with about half of
the feature tests here, so it seems to be following a reasonable set of
rules. I've therefore applied VS Code's formatting fixes to all the
feature files so they're now consistent, and I'll be able to avoid
future noisy diffs caused by my editor.