Switch on the the feature flag for Release 1.3
This force-enables the flag for the Whitehall transition to Design
System release 1.3.

This is an easy way to make the release available to all users quickly,
but should be followed up by the removal of `preview_design_system?` for
all pages currently flagged as `next_release: true` since they'll no
longer be 'unreleased' changes.