Update controller tests to use the GOV.UK Design System
A bunch of tests have started failing now we're ensuring we render
pages in the upcoming 1.3 release render the design system
implementation.

These tests fall into 1 of 2 categories.

1. They were not duplicated and only test the Bootstrap implementation
2. They were duplicated, one with the Preview design system permission,
and one without.

In the first instance they need to be updated to use the GOV.UK Design
System implementation. And in the second, we can remove the Bootstrap
spec.