Revert consultation response attachments to use the Bootstrap implementation
We recently ported the attachment index page to use a GOV.UK Design
System implementation. When we renamed the Bootstrap partial
to attachments_legacy.html.erb, we missed updating the name of the
partial in the response show page.

This has led to the Design System partial being used on a page
that otherwise has not been ported to use the Design System. Breaking
the styling and functionality.