Add ADR and documentation for View Components
We've come to a consensus on how we want to proceed using View
Components. The discussion surrounding this can be found in this
Github Issue https://github.com/alphagov/whitehall/issues/6954.

This adds an ADR outlining this, and a view components doc, which
outlines when to use them and the file structure they should
follow.