Ensure base errors render as text not link
In some situations an object cannot be saved due to top
level errors which are not based on the users input.

For example, you cannot withdraw a published edition if it has a
draft. In this situation an error is added to the base attr of the
object.

These do not link to specific inputs so we shouldn't render them
as links.