Make attachment index page object agnostic
At the moment, this view assumes an edition will be passed in. However,
other objects have attachments so we should make the view as object
agnostic as possible.

As far as i can tell there are three types of attachables that use this
view.

1. editions
2. responses
3. policy groups

I've ensured the links work for all 3 types and have added confirm_destroy
endpoints for responses and policy groups. And a reorder endpoint for
responses.

Policy groups attachments cannot be reordered so one is not required.