Fix flakey scheduled publishing test
When running tests in parallell the scheduled publishing worker test has
a tendency to randomly fail. This seems to be because old editions are
not properly discarded or are running at the same time which causes the
test to fail.

To fix this we have stubbed out the sidkiq jobs so they won't interfere
with eachother.

Trello card: https://trello.com/c/DhRIp7Nu/830-flaky-scheduled-publishing-worker-tests-in-whitehall