Fix flakey scheduled publishing test
When running tests in parallell the scheduled publishing worker test has
a tendency to randomly fail. This seems to be because old editions are
not properly discarded or are running at the same time which causes the
test to fail.

To fix this we have set up the tests to run in a different queue for
each parallel process that runs. This should prevent the other queues
from interfering with the current running test.

We have also move away from using the redis namespace config as it is
being removed in the next major version.

Trello card: https://trello.com/c/DhRIp7Nu/830-flaky-scheduled-publishing-worker-tests-in-whitehall

Update test/support/sidekiq_test_helpers.rb

Co-authored-by: Ollie Treend <ollie.treend@digital.cabinet-office.gov.uk>