Update fact checking edit and show page based on design feedback
This updates the fact checking edit page to use the inset text,
rather than the notification banner for the instructions provided to the
SME.

It also moves it into the left 2/3 of the page rather than the right 1/3
above the textarea.

It also updates the H1 and flash message on the fact checking show page.

Both of these changes are based on Design feedback.