Change method of selecting schema test cases
We are currently running schema tests by grepping the test names for a
specific string. However there are tests for valid schemas that do not
have a name in this format that we are currently missing.

By running all the Publishing API presenter tests (and a validation test
for the hardcoded finders), we will ensure no tests are missed.

This adds an extra (approximately) 15 seconds to the test run, but
ensures we have full test coverage.