Use correct URL for International Delegation pages
We currently have two almost identical pages for International
Delegations: `/world/slug` and `/world/slug/news`.

We are deprecating the news version of this page, so need to update the
presenter to present the correct URL.

Once this has been merged, all International Delegation pages will need
to be republished.