Stop World Locations being published
We were previously publishing the World Location page to `/world/slug`
for International Delegation pages.

This is no longer needed as we publish the World Location News page to
the same base path.

If we don't stop the World Location being published, we could end up
with a conflict where the World Location content item may overwrite the
World Location News content item for the International Delegation.

This does not affect non-International Delegation World Locations, as
they are published by Content Tagger.

The ideal solution would be to separate International Delegations into
their own models, seperate to World Locations, but this is out of scope
of the current work.