Add FactCheckingTabComponent and render in fact checking tab
This adds a fact checking tab which handles rendering content on the
fact checking tab.

It ensures that fact check requests are separated into sections based on
whether they have responses or are awaiting a response.

It ensures that form fields for requesting a new fact check are
only rendered when the `send_request_section` arg is set to `true`
on initialisation.